pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "camtrace"
version = "0.1.0"
description = "CamTrace network flow enrichment (PTR/ASN/Geo via local MaxMind DBs)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ty Carruth" }]

# Pin with conservative upper-bounds to reduce breaking changes; lock exact versions in requirements.txt via pip-tools.
dependencies = [
  "pydantic>=2.6,<3",
  "dnspython>=2.6,<3",
  "geoip2>=4.8,<5",
]

# Make dotenv optional (dev/local only)
[project.optional-dependencies]
dev = [
  "python-dotenv>=1.0,<2",
  "ruff>=0.6.8,<1",
  "bandit>=1.7,<2",
  "pip-audit>=2.7,<3",
  "pre-commit>=3.8,<4",
  "detect-secrets>=1.5,<2",
]
test = [
  "pytest>=8.3,<9",
  "pytest-cov>=5.0,<6",
]
# (add a "cli" extra if you later pull in click/typer)

[project.scripts]
camtrace = "camtrace.cli:main"
camtrace-capture = "camtrace.capture_cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["camtrace"]

[tool.setuptools.package-data]
camtrace = ["bin/pcap_to_camtrace.sh"]

# --- Tooling (security/quality) ---
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["dist", "build", ".venv"]

[tool.ruff.lint]
extend-select = ["E", "F", "I", "UP", "B"]

[tool.bandit]
skips = ["B101"]  # assert usage; adjust as needed

